+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+pnet-data-api 2.0.3
+====================
+
+* Add autoComplete to applications.
+
 pnet-data-api 2.0.2
 ===================
 
@@ -17,6 +22,11 @@ pnet-data-api 2.0.0
   older Spring Boot version (since the Data API does not depend on it, nor does is use any Jakarta packages). You still
   have to use up-to-date libraries like Jackson 2.14 or greater, Spring-Web 6.0 or greater (if using the Spring
   connector) and Java 17 or greater.
+
+pnet-data-api 1.21.3
+====================
+
+* Add autoComplete to applications.
 
 pnet-data-api 1.21.2
 ====================